refactor: split into modular package structure
Refs: #3
