+++ gh_audit/__init__.py
@@ -1,0 +1,1 @@
+# This file makes gh_audit a package.
