+++ gh_audit/rules/__init__.py
@@ -1,0 +1,5 @@
+from .dependabot import *  # noqa: F401, F403
+from .general import *  # noqa: F401, F403
+from .github_actions import *  # noqa: F401, F403
+from .nix import *  # noqa: F401, F403
+from .python import *  # noqa: F401, F403
