+++ pyproject.toml
@@ -13,9 +13,15 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
 ]
 
-[project.entry-points.console_scripts]
-gh-audit = "gh_audit:main"
+[dependency-groups]
+dev = [
+    "mypy>=1.0.0,<2.0",
+    "ruff>=0.4.0",
+    "types-PyYAML>=6.0.1",
+]
 
+[project.scripts]
+gh-audit = "gh_audit.cli:main"
 
 [tool.ruff.lint]
 extend-select = ["I", "UP"]
@@ -23,9 +29,6 @@ extend-select = ["I", "UP"]
 [tool.mypy]
 strict = true
 
-[dependency-groups]
-dev = [
-    "mypy>=1.0.0,<2.0",
-    "ruff>=0.4.0",
-    "types-PyYAML>=6.0.1",
-]
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
